Practice for open world gaming
I started of trying to program an entire game in one sitting. Needless to say, that didn't work. My current strategy is to simply go with what works. It may not be pretty, but it will work and I will actually learn something. These files each deal with a different aspect of the open world role playing game.